liveliness: package to represent backend peers
For any backend server to primary connection, represent
it with a convenience model that will be used by the frontend
when proxying requests to each backend.

Periodically, per cycle, the primary will perform a /ping to
each backend server and then figure out who is alive and who isn't,
then proxy traffic only to those alive until the next cycle
for which the process will repeat.