load-distribution: implement proxying to multiple backends
Integrated package ./lively so that the frontend server
can distribute load to multiple backends, but periodically
it can also figure out which servers are live and which aren't
thereby maintaining liveliness and directing to traffic to
those that can receive it, per cycle.
For starters during a cycle, the load is distributed to
live backends in round robin style. However, at the end of every
cycle, the live backends list is recompiled, shuffled(for some fairness)
and then the round robin distribution starts again.