cmd/frontender: CLI started + hooked up with lively
* Made a CLI that can now be runnable by anyone
* Hooked up frontender with the liveliness check package
and performed an end to end test
* Made an exception for easy compatibility with older(and most)
systems because not everyone is going to want to have to go change
their code base to include a /ping route just to use this server.
The exception now is that if lively encounters a 404 Not Found,
we know the backend server is alive and can therefore mark it
as lively since it actually responded.