lively: ensure ping closes body irrespective of status code
Ensure that ping closes its response body irrespective of
status code. I noticed this while investigating a test
deployment using frontender as the frontend server on
Google Compute Engine and noticed a bunch of leaks.
Added a test to ensure that we always close the response body
irrespective of status code, as long as it is non-nil.